+++ src/Resources/app/administration/src/module/const.js
@@ -1,7 +1,4 @@
-// export const LENGOW_URL = 'https://my.lengow.io';
-export const LENGOW_URL = 'https://my.lengow.net';
-
-export const MODULE_VERSION = '1.1.2';
+export const MODULE_VERSION = '1.1.3';
 
 export const ORDER_LENGOW_STATES = {
     accepted: 'accepted',