+++ src/Resources/app/administration/src/module/lengow-connector/component/lgw-footer/index.js
@@ -1,24 +1,54 @@
 import template from './lgw-footer.html.twig';
 import './lgw-footer.css';
-import { LENGOW_URL, MODULE_VERSION } from '../../../const';
+import { MODULE_VERSION } from '../../../const';
 
-const { Component } = Shopware;
+const {
+    Component,
+    Data: { Criteria }
+} = Shopware;
 
 Component.register('lgw-footer', {
     template,
 
+    inject: ['repositoryFactory'],
+
     data() {
         return {
-            lengowUrl: LENGOW_URL,
+            lengowEnvironmentUrl: 'https://my.lengow.net',
             moduleVersion: MODULE_VERSION,
             currentYear: new Date().getFullYear(),
             preprod: false
         };
     },
 
-    created() {
-        if (this.lengowUrl === 'https://my.lengow.net') {
-            this.preprod = true;
+    computed: {
+        lengowConfigRepository() {
+            return this.repositoryFactory.create('lengow_settings');
         }
+    },
+
+    created() {
+        this.loadEnvironmentUrl();
+    },
+
+    methods: {
+        loadEnvironmentUrl() {
+            const lengowConfigCriteria = new Criteria();
+            lengowConfigCriteria.addFilter(Criteria.equals('name', 'lengowEnvironmentUrl'));
+            this.lengowConfigRepository.search(lengowConfigCriteria, Shopware.Context.api).then(result => {
+                if (result.total > 0) {
+                    this.lengowEnvironmentUrl = 'https://my.lengow' + result[0].value;
+                    this.checkPreprod();
+                }
+            });
+        },
+
+        checkPreprod() {
+            if (this.lengowEnvironmentUrl === 'https://my.lengow.net') {
+                this.preprod = true;
+            } else {
+                this.preprod = false;
+            }
+        },
     }
 });