+++ src/Resources/app/administration/src/module/lengow-connector/view/lgw-dashboard-free-trial/index.js
@@ -1,18 +1,30 @@
 import template from './lgw-dashboard-free-trial.html.twig';
 import './lgw-dashboard-free-trial.scss';
-import { LENGOW_URL } from '../../../const';
 
-const { Component } = Shopware;
+const {
+    Component,
+    Data: { Criteria }
+} = Shopware;
 
 Component.register('lgw-dashboard-free-trial', {
     template,
 
-    inject: ['LengowConnectorSyncService'],
+    inject: ['LengowConnectorSyncService', 'repositoryFactory'],
 
     data() {
         return {
-            lengow_url: LENGOW_URL
+            lengowUrl: 'https://my.lengow.net'
         };
+    },
+
+    computed: {
+        lengowConfigRepository() {
+            return this.repositoryFactory.create('lengow_settings');
+        }
+    },
+
+    created() {
+        this.loadEnvironmentUrl();
     },
 
     methods: {
@@ -20,6 +32,17 @@ Component.register('lgw-dashboard-free-trial', {
             this.LengowConnectorSyncService.getAccountStatus(true).then(result => {
                 if (result.success) {
                     window.location.reload();
+                }
+            });
+        },
+
+        loadEnvironmentUrl() {
+            const lengowConfigCriteria = new Criteria();
+            lengowConfigCriteria.addFilter(Criteria.equals('name', 'lengowEnvironmentUrl'));
+            this.lengowConfigRepository.search(lengowConfigCriteria, Shopware.Context.api).then(result => {
+                if (result.total > 0) {
+                    this.lengowUrl = 'https://my.lengow' + result[0].value;
+                    this.createdComponent();
                 }
             });
         }