+++ src/Resources/app/administration/src/module/lengow-connector/view/lgw-dashboard-home/index.js
@@ -1,23 +1,33 @@
 import template from './lgw-dashboard-home.html.twig';
 import './lgw-dashboard-home.scss';
-import { LENGOW_URL } from '../../../const';
 
-const { Component } = Shopware;
+const {
+    Component,
+    Data: { Criteria }
+} = Shopware;
+
 
 Component.register('lgw-dashboard-home', {
     template,
 
-    inject: ['LengowConnectorSyncService'],
+    inject: ['repositoryFactory','LengowConnectorSyncService'],
 
     data() {
         return {
-            lengow_url: LENGOW_URL,
+            lengowEnvironmentUrl: 'https://my.lengow.net',
             helpCenterLink: ''
         };
     },
 
+    computed: {
+        lengowConfigRepository() {
+            return this.repositoryFactory.create('lengow_settings');
+        }
+    },
+
     created() {
         this.createdComponent();
+        this.loadEnvironmentUrl();
     },
 
     methods: {
@@ -27,6 +37,16 @@ Component.register('lgw-dashboard-home', {
                     this.helpCenterLink = result.links.help_center;
                 }
             });
+        },
+        loadEnvironmentUrl() {
+            const lengowConfigCriteria = new Criteria();
+            lengowConfigCriteria.addFilter(Criteria.equals('name', 'lengowEnvironmentUrl'));
+            this.lengowConfigRepository.search(lengowConfigCriteria, Shopware.Context.api).then(result => {
+                if (result.total > 0) {
+                    this.lengowEnvironmentUrl = 'https://my.lengow' + result[0].value;
+                }
+            });
         }
+
     }
 });